requirements.txt
# requirements.txt for Ethical Hacking Dashboard (Streamlit)
streamlit==1.39.0
opencv-python==4.10.0.84
requests==2.32.3
speechrecognition==3.10.4
pyaudio==0.2.14
scapy==2.5.0
python-nmap==0.7.1
psutil==6.0.0
cryptography==43.0.1
pynput==1.7.7
beautifulsoup4==4.12.3